对文章的权限控制

+++ app/Http/Controllers/PostController.php
@@ -27,7 +27,9 @@ class PostController extends Controller
             'content'=>'required|string|min:10',
         ]);
         //2.逻辑
-        $post=Post::create(\request(['title','content']));
+        $user_id = \Auth::id();
+        $params = array_merge(request(['title','content']),compact('user_id'));
+        $post=Post::create($params);
         //3.渲染
         return redirect("/posts");
 
@@ -38,14 +40,13 @@ class PostController extends Controller
     }
     //修改文章接口
     public function update(Post $post){
-
-        //TODO 用户权限认证
-
         //1.验证
         $this->validate(request(),[
             'title'=>'required|string|max:100|min:5',
             'content'=>'required|string|min:10',
         ]);
+        //用户权限认证
+        $this->authorize('update',$post);
         //2.逻辑
         $post->title = \request('title');
         $post->content = \request('content');
@@ -55,7 +56,8 @@ class PostController extends Controller
     }
     //删除文章接口
     public function delete(Post $post){
-        //TODO 用户权限认证
+        //用户权限认证
+        $this->authorize('delete',$post);
         $post->delete();
 
         return redirect("/posts");
